+++ live/livekit/lib/component/float_window/global_float_window_manager.dart
@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:tencent_live_uikit/common/widget/float_window/float_window_mode.dart';
 
 import 'package:tencent_live_uikit/common/platform/rtc_live_tuikit_platform_interface.dart';
+import 'package:tencent_live_uikit/common/widget/global.dart';
 import 'global_float_window_state.dart';
 
 /// Two scenarios:
@@ -64,6 +65,7 @@ class OverlayManager {
   VoidCallback? _switchToFullScreenMode;
 
   void showOverlayEntry(OverlayEntry overlayEntry) {
+    Global.secondaryNavigatorKey.currentState!.overlay!.insert(overlayEntry);
     _overlayEntry = overlayEntry;
   }