+++ live/livekit/lib/component/float_window/global_float_window_state.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import '../../common/widget/float_window/float_window_mode.dart';
 
 class GlobalFloatWindowState {
-  bool enableFloatWindowFeature = true;
+  bool enableFloatWindowFeature = false;
   final ValueListenable<FloatWindowMode> floatWindowMode = ValueNotifier(FloatWindowMode.none);
   final ValueListenable<String> roomId = ValueNotifier("");
   final ValueListenable<String> ownerId = ValueNotifier("");