+++ live/livekit/lib/voice_room/widget/panel/live_background_select_widget.dart
@@ -111,6 +111,7 @@ class _LiveBackgroundSelectPanelWidgetState extends State<LiveBackgroundSelectPa
                                   return Image.asset(
                                     LiveImages.defaultBackground,
                                     package: Constants.pluginName,
+                                    fit: BoxFit.cover,
                                   );
                                 }),
                           );